Add hook to lab plate reader
