+++ labop/strings.py
@@ -1,0 +1,13 @@
+class Strings:
+    ALIQUOT = "aliquot"
+    SAMPLE = "sample"
+    METADATA= "metadata"
+    DATA = "data"
+    XARRAY = "xarray"
+    JSON = "json"
+    MASK = "mask"
+    MEASUREMENT = "measurement"
+    CONTENTS = "contents"
+    SOURCE = "source"
+
+
